\name{DefectData-package}
\alias{DefectData-package}
\alias{DefectData}
\docType{package}
\title{
An R package of Defect Prediction Datasets for Software Engineering Research
}
\description{
An R package of Defect Prediction Datasets for Software Engineering Research
}
\details{
\tabular{ll}{
Package: \tab DefectData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-27\cr
License: \tab MIT\cr
}

}
\author{
Chakkrit (Kla) Tantithamthavorn

Maintainer: Chakkrit (Kla) Tantithamthavorn <kla@chakkrit.com>
}
\references{
-
}
-
\keyword{ package }
\seealso{
-
 \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
# What data are available in this R package
# listData

# Loading a dataset by specifying a system name
# loadData("JM1")
}
