% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listData.R
\name{listData}
\alias{listData}
\title{List of defect datasets}
\format{A data frame with 101 rows and 7 variables}
\source{
\url{http://www.openscience.us/repo}
}
\description{
This command lists all defect datasets that are available in this R package
}
\details{
\itemize{
  \item system. the system name.
  \item corpus. the corpus of wher it comes from.
  \item DefectiveRatio. the defective ratio of the data.
  \item Modules. the number of modules (observations).
  \item Defective. the number of defective modules.
  \item Predictors. the number of predictors (independent variable).
  \item EPV. the number of event per variables (EPV).
}
}

