% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{A Load Defect Data function}
\usage{
loadData(system)
}
\arguments{
\item{system}{system name}
}
\value{
an object of data
}
\description{
This function allows you to load defect data in software engineering research
}
\examples{
data <- loadData("jdt")
}

