% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listData.R
\name{listData}
\alias{listData}
\title{A list defect data function}
\usage{
listData()
}
\description{
This function that allows you to list all defect datasets that are available in this R package
}
\examples{
listData()
}

